Update keymap.json to match adv360.keymap
The web UI seems to want to keep the two in sync.  ZMK doesn't generally
seem to require this JSON file at all, but chatter on the Discord
suggests that Kinesis has some nonstandard mods on top of regular ZMK,
and maybe not updating both is why my last firmware didn't work?